requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dependency-injector==4.37.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==4.0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.2.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==4.0.3
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.10.3
packaging==24.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
pytest==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.15.0
uvloop==0.19.0
vine==5.1.0
watchgod==0.8.2
wcwidth==0.2.13
websockets==12.0
